Cleaned things up a bit more, added a type for building a dynamic collision mesh. Not sure if I'm totally happy with this, but it works for now.

+++ LibSm64Sharp/src/ISm64CollisionMesh.cs
@@ -7,7 +7,7 @@
     TSelf AddTriangle(
         Sm64SurfaceType surfaceType,
         Sm64TerrainType terrainType,
-        IReadOnlyList<(float x, float y, float z)> vertices);
+        IReadOnlyList<(short x, short y, short z)> vertices);
   }
 
   public interface ISm64CollisionMesh {
